Added new functions for displaying scrambles + more
Added functions for displaying the current scramble and to display the scrambled cube visually.
Copied the functions within cube.py to the main program because with importing these functions some issues with global variables arose. Mainly I couldn't easily mutate the faces of the cube because they were unavailable to the functions in the other file. This could probably be fixed later to make the code more readable but at the moment things work as they do.
Also updated the main loop to display a new scramble every time a time is recorded.

Added functions: drawFace, drawCube, displayScramble, getColour (also added new colour variables to make this function work)